Move ForgotPasswordModal to components dir
